Add a JSON endpoint for the most recent posts
Why:

This change creates a JSON endpoint for the most recent blog posts.

We need to render the most recent posts on the homepage. The homepage
is a static page, built on middleman, and adding a back-end to list blog
posts seems overkill. The idea is to make a request on the browser for the
most recent posts and add them dynamically to the homepage.

Notes:

I'm not very confident about duplicating information about the picture's
name, I guess on the model would make more sense to avoid duplication.
But on the other hand, it would add presentation details to the model.

This change addresses the need by:

* Adding AMS.
* Adding serializers for posts and users.
* Adding a JSON endpoint.