Adds dedicated hero model (part 1 of 2)
Previously, hero images were stored in the same model as the rest of the
post images.

I want to have more control over them, and process them in a different
way, so I'm migrating them to a dedicated Hero model

In this first PR, I create the Hero model and uploader, and add a rake
task which I will run in production to generate Hero instances for
existing posts. Posts are still referring to the original PostImage
record though, so no actual changes are happening

A second PR (after I make sure the data migration worked) will make the
switch, and delete all old occurences of PostImage records