Changes default branch name so as not to depend on post title
Whenever a title changes, it's corresponding PR is no longer found, and
a new one is created.
This changes the behaviour so that the branch name no longer depends on
the title, and on the author's name and post ID instead, since those are
not expected to change.

The previous behaviour was deprecated and not removed, so that existing
PRs still work